+++ src/domain/exploration/creature_gift_claim_result.gd
@@ -1,0 +1,7 @@
+class_name CreatureGiftClaimResult
+extends RefCounted
+
+var success: bool = false
+var reason: StringName = &""
+var creature: CreatureInstance
+var destination: StringName = &""